score-service/requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
google-generativeai
sentence-transformers
kafka-python
